refactor: extract buildRecordWithAllMorphObjectIdsToNull utility
Address PR review feedback:
- Rename allNullRecordInput to recordWithAllMorphObjectIdsToNull for clarity
- Extract reducer logic into a reusable utility function
